Add basic results plotting
This commit adds the basic skeleton for plotting the power output with time. Will still need to tidy up the plots to make them look pretty etc. Also changed the name of module "io.py" to "user_interface.py" because the io name was conflicting with a built in python module.