attempts to add some tests
Attempted to add a test suite but encountered a strange error resulting in a Python fatal error/segmentation fault. Think it may be due to some interaction between Pytest and the Orekit Java VM as when the error occurs it also spits out a log file with a Java error trace. Error seems to occur on the first use of Orekit AbsoluteDate in conftest.py